Add support for multiple schemata in .avsc files
Not all .avsc files are a singular schema.  In practice, the .avsc files
we're working with using PFBs are in the format of a list of schemata.

To support this, the multi-avsc2avdl.sh script will split an .avsc file
into multiple .avsc files, each of which contains a single schema.

The avsc2avdl Java program is modified and used to convert each schema
individually.